Module Dune_glob.V1

Simple glob support library.

type t

val empty : t

A glob that matches nothing

val universal : t

A glob that matches anything (including the strings starting with a ".")

val test : t -> string -> bool

Tests if string matches the glob.

val to_string : t -> string

Returns textual representation of a glob.

val of_string : string -> t

Converts string to glob. Throws Invalid_argument exception if string is not a valid glob.

val of_string_result : string -> (t, int * string) Stdune.result

val to_dyn : t -> Dyn.t

val of_string_exn : Stdune.Loc.t -> string -> t

val compare : t -> t -> Stdune.Ordering.t

val hash : t -> int

val matching_extensions : Stdune.Filename.Extension.t list -> t

matching_extensions xs return a glob that will match any of the extensions in xs